+++ lib/gtp/v2/path.h
@@ -17,42 +17,36 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-#ifndef OGS_GTP_H
-#define OGS_GTP_H
+#if !defined(OGS_GTP_INSIDE) && !defined(OGS_GTP_COMPILATION)
+#error "This header cannot be included directly."
+#endif
 
-#include "ogs-core.h"
-#include "ipfw/ogs-ipfw.h"
-#include "ogs-app.h"
-#include "ogs-nas-common.h"
-
-#define OGS_GTPV1_U_UDP_PORT            2152
-#define OGS_GTPV2_C_UDP_PORT            2123
-
-#define OGS_GTP_INSIDE
-
-#include "gtp/v2/message.h"
-#include "gtp/v2/types.h"
-#include "gtp/v2/conv.h"
-#include "gtp/context.h"
-#include "gtp/v2/build.h"
-#include "gtp/v2/path.h"
-#include "gtp/path.h"
-#include "gtp/xact.h"
-#include "gtp/util.h"
+#ifndef OGS_GTP2_PATH_H
+#define OGS_GTP2_PATH_H
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-#undef OGS_GTP_INSIDE
+typedef struct ogs_gtp_xact_s ogs_gtp_xact_t;
 
-extern int __ogs_gtp_domain;
 
-#undef OGS_LOG_DOMAIN
-#define OGS_LOG_DOMAIN __ogs_gtp_domain
+int ogs_gtp_send_user_plane(
+        ogs_gtp_node_t *gnode,
+        ogs_gtp_header_t *gtp_hdesc, ogs_gtp_extension_header_t *ext_hdesc,
+        ogs_pkbuf_t *pkbuf);
+
+ogs_pkbuf_t *ogs_gtp_handle_echo_req(ogs_pkbuf_t *pkt);
+void ogs_gtp_send_error_message(
+        ogs_gtp_xact_t *xact, uint32_t teid, uint8_t type, uint8_t cause_value);
+
+void ogs_gtp_send_echo_request(
+        ogs_gtp_node_t *gnode, uint8_t recovery, uint8_t features);
+void ogs_gtp_send_echo_response(ogs_gtp_xact_t *xact,
+        uint8_t recovery, uint8_t features);
 
 #ifdef __cplusplus
 }
 #endif
 
-#endif /* OGS_GTP_H */
+#endif /* OGS_GTP2_PATH_H */
